requirements-dev.txt
# Dependencias para desarrollo local (SQLite)
# PostgreSQL se configurará automáticamente en producción (Railway)

# Dependencia necesaria para Python 3.13
setuptools>=65.5.0

Django==5.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-filter==23.3
django-redis==5.4.0
django-debug-toolbar==6.0.0
django-extensions==4.1
python-decouple==3.8
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
coverage==7.3.4
drf-spectacular==0.27.1
